move nemo_ckpt_convert into the converter folder
